requirements.txt
numpy
pandas
matplotlib
kaggle
torch
torchvision
torchaudio
pytorch-lightning
wandb
tqdm
Pillow
segmentation-models-pytorch